Add K means clustering
This is via the kmlocal code from
http://www.cs.umd.edu/~mount/Projects/KMeans/. We include this code in
vendor and build it out as a static archive, which then will be linked
into our final products.

When doing the clustering we statically set some of the clustering
parameters. These parameters are pseudo-magical and, for now, I'm not
going to bother trying to figure out what the mean. Space for future
research and tweaking, but for now I just want something that works.