Don't search for OpemMP include when building gcc
We use openmp and gcc on Travis, which has openmp stuff installed to
system directories. The header search is really a OSX specific hack for
how homebrew installs iomp headers, so we should do an os check or
something similar instead of a compiler check. However, for now I'm just
getting stuff working.